ecommerce-ai/backend/requirements.txt
flask>=2.2.0
flask-cors==3.0.10
tensorflow-cpu==2.16.1
numpy>=1.24.3
Pillow==10.0.0
python-dotenv==1.0.0 